safari-600.7-branch/LayoutTests/js/dfg-to-string-bad-valueOf-expected.txt

Tests that we do ToString conversions correctly when String.prototype.valueOf is not what we wanted.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS "" + foo("hello") is "42"
PASS successfullyParsed is true

TEST COMPLETE
